fix bugs

+++ libasync/src/ChaseLevDeque.h
@@ -42,7 +42,7 @@ class ChaseLevDeque
 {
 public:
     // ChaseLevDeque takes log2(capacity) as argument to guarantee that its capacity is power of 2. 
-    ChaseLevDeque(int log2capacity=13)  : top(1<<(log2capacity-1)), bottom(1<<(log2capacity-1)),
+    ChaseLevDeque(int log2capacity=13)  : top((uint64_t)1<<(log2capacity-1)), bottom((uint64_t)1<<(log2capacity-1)),
                                         array(new SharedPointerArray(1<<log2capacity))
     {}
     struct SharedPointerArray{
